DevIdea/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "numpy>=2.2.3",
    "openai>=1.65.4",
    "pandas>=2.2.3",
    "passlib>=1.7.4",
    "plotly>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.38",
    "streamlit>=1.43.1",
    "trafilatura>=2.0.0",
]
